tests(infra): split infra tests ownership
