+++ x-pack/test/functional/apps/infra/index.ts
@@ -23,12 +23,12 @@ export default ({ loadTestFile }: FtrProviderContext) => {
     });
 
     describe('Logs UI', function () {
-      loadTestFile(require.resolve('./log_entry_categories_tab'));
-      loadTestFile(require.resolve('./log_entry_rate_tab'));
-      loadTestFile(require.resolve('./logs_source_configuration'));
-      loadTestFile(require.resolve('./log_stream_date_nano'));
-      loadTestFile(require.resolve('./link_to'));
-      loadTestFile(require.resolve('./log_stream'));
+      loadTestFile(require.resolve('./logs/log_entry_categories_tab'));
+      loadTestFile(require.resolve('./logs/log_entry_rate_tab'));
+      loadTestFile(require.resolve('./logs/logs_source_configuration'));
+      loadTestFile(require.resolve('./logs/log_stream_date_nano'));
+      loadTestFile(require.resolve('./logs/link_to'));
+      loadTestFile(require.resolve('./logs/log_stream'));
     });
   });
 };
