+++ x-pack/test/functional/apps/infra/logs/log_entry_rate_tab.ts
@@ -7,7 +7,7 @@
 
 import expect from '@kbn/expect';
 
-import { FtrProviderContext } from '../../ftr_provider_context';
+import { FtrProviderContext } from '../../../ftr_provider_context';
 
 export default ({ getService }: FtrProviderContext) => {
   const logsUi = getService('logsUi');
